backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=24.2
Pillow>=10.0
psycopg2-binary>=2.9
gunicorn>=21.2
